finish (sound added, coin data type change to Long for big number, fix some layout)

+++ app/src/main/java/id/kotlin/osom/Auth/dataProfile.kt
@@ -3,6 +3,6 @@ package id.kotlin.osom.Auth
 data class dataProfile(
     val id : String? = null,
     val username : String,
-    val coin : Int? = null,
+    val coin : Long? = null,
     val email : String,
 )
