+++ app/src/main/java/id/kotlin/osom/CoinOsomAPI/dataOsom.kt
@@ -2,5 +2,5 @@ package id.kotlin.osom.CoinOsomAPI
 
 data class dataOsom(
     val id : String? =null,
-    val coin : Int,
+    val coin : Long,
 )
